Create TickerPriceScraper.py
I tried to write a simple program that finds the current price of a given ticker symbol. However, alpaca was difficult to scrape from, most likely because the prices are constantly changing and what  I really wanted to show as the output is the high and low for that day. This is only visible via stock charts. I sort of gave up.