Attempt at real time data retrievement
Experiments with json and a websocket, but does not work, was hoping it could be looked at. Was thinking there was a problem with wscat and Node versions. Wscat allows your websocket to connect. It is a module that can be installed globally if Node js is installed, which I did.